add code for the top claw and refactor references to the old claw into a new class called bottomClaw
