+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/auton/TimeAuton.java
@@ -4,7 +4,7 @@ import com.acmerobotics.dashboard.config.Config;
 import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 import com.qualcomm.robotcore.util.ElapsedTime;
-import org.firstinspires.ftc.teamcode.utilities.Claw;
+import org.firstinspires.ftc.teamcode.utilities.topClaw;
 
 import org.firstinspires.ftc.teamcode.utilities.SimpleMecanumDrive;
 
@@ -13,7 +13,7 @@ import org.firstinspires.ftc.teamcode.utilities.SimpleMecanumDrive;
 public class TimeAuton extends LinearOpMode {
     public static int slideAmount = 1000;
 
-    private Claw claw;
+    private topClaw claw;
 
     @Override
     public void runOpMode() throws InterruptedException {
