+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/teleop/Teleop.java
@@ -6,7 +6,7 @@ import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
 
 import org.firstinspires.ftc.teamcode.utilities.Arm;
 import org.firstinspires.ftc.teamcode.utilities.SimpleMecanumDrive;
-import org.firstinspires.ftc.teamcode.utilities.Claw;
+import org.firstinspires.ftc.teamcode.utilities.topClaw;
 import org.firstinspires.ftc.teamcode.utilities.Slides;
 import org.firstinspires.ftc.teamcode.utilities.SlideState;
 import org.firstinspires.ftc.teamcode.utilities.PullUp;
@@ -26,7 +26,7 @@ public class Teleop extends OpMode {
     private TeleopState teleopState = TeleopState.INIT;
 
     private SimpleMecanumDrive drive;
-    private Claw claw;
+    private topClaw claw;
     private Slides slides;
     private Arm arm;
     private PullUp pullUp;
@@ -48,7 +48,7 @@ public class Teleop extends OpMode {
     public void init() {
         // initialize util classes for hardware
         drive = new SimpleMecanumDrive(hardwareMap);
-        claw = new Claw(hardwareMap);
+        claw = new topClaw(hardwareMap);
         slides = new Slides(hardwareMap);
         arm = new Arm(hardwareMap);
         pullUp = new PullUp(hardwareMap);
